Move providers from `app` to `routes`
