+++ src/node/routes/static.ts
